+++ KenStockBallot/KenStockBallot/Classes/ViewController/KSBMainCalculateVC.m
@@ -271,7 +271,7 @@ static const int cellOffX = 0;
     KSBStockInfo *info = _dataArray[indexPath.row];
     NSArray *array = @[info.stockName, info.stockJiaoYS,
                        [NSString stringWithFormat:@"%.2f%@", info.stockPrice, KenLocal(@"edit_yuan")],
-                       [NSString stringWithFormat:@"%d%@", info.stockBuyMax, KenLocal(@"edit_gu")],
+                       [NSString stringWithFormat:@"%d%@", (int)info.stockBuyMax, KenLocal(@"edit_gu")],
                        [NSString stringWithFormat:@"%.2f%%", info.stockBallot]];
     float width = (kGSize.width - cellOffX) / [array count];
     for (int i = 0; i < [array count]; i++) {
